Improve the way filters are dependents
Filters are complex: they can filter on both a flag (UNSEEN, etc) or a
range (sort by date from..to). To make Ui more clear, we disable a
filter when activate another, event if they're not excludables (e.g.
date range and attachments). This commit attach this logic to the
component state, so Ui should only take care about the buttons state
and the sorting will do it lonely, based on the current state.

It also introduces the availability for the parent component (i.e. the
toolbar) to overrides the state of its child component (the range date
picker) so the active sate can be setted internally by the date picker
component, or externally (i.e. for de-activation) by its parent.