Fix bare newlines when sending message
When sending a message, some IMAP servers'll complain about _bare newlines_ when
the message is saved to the `sent` folder after smtp submission. This can be due
to RFC enforcements that forbid bare newlines after headers.

Our editor always starts with a `\n` character, and it may produces the bare
newlines warning from IMAP server as it follows directly the message header.

This commit trim the content of the message to remove this useless newline
first-char and try to fix the warning.