[feat] force autodiscover reset
When we try to discover an erroneous account, we disable autodiscover when
receiving autodiscover error. This commit introduces a new feature that, when
correcting the email/login field, and if the servers fields are always empty,
we transparently re-enable the autodiscover feature.

It addresses the following use-case: I fill the email/login field but make a
typo in my email address. The autodiscover so failed, and I fix the typo. I
don't want to manually fill the advanced params but retry an autodiscover. Only
exception is if I already started to fill servers fields, we then assume we want
to set the account manually and leave the autodiscover feature disabled.