[perf] throttle XHR requests to prevent gateway timeouts
Sometimes, huge bench of requests struggle server side app, which may it hanging
on, and occurs gateway timeouts for next upcoming request as the server is still
idle.
This commit throttle XHR requests in a given framerate to ensure the server
won't be overload and can always respond to a request. We don't throttle every
requests but only :
- all GET requests (especially bulk fetch requests)
- huge or frequent PUT/POST requests that may struggle the server